Add curl-noise advection and in-engine video recording
- Curl noise: advect particles by incompressible velocity from analytic
  fbm/perlin curl; density sourcing for colored trails (NoiseSourcingCS)
- Fix boundary teleport: drop frac() in the finite-difference stencil and
  correct dx axis in getCurl
- Backbuffer recording: fixed-dt offline capture piped directly to ffmpeg
  for mp4 (decoupled capture/output fps), reusing readback->RGBA path

+++ SEngine/StaticMesh.cpp
@@ -45,5 +45,5 @@ void StaticMesh::Render(ID3D12GraphicsCommandList* commandList)
 void StaticMesh::RenderPoints(ID3D12GraphicsCommandList* c)
 {
     c->IASetPrimitiveTopology(D3D_PRIMITIVE_TOPOLOGY_POINTLIST);
-    c->DrawInstanced(m_indexCounts[0], 1, 0, 0);
+    c->DrawInstanced((UINT)m_indexCounts[0], 1, 0, 0);
 }
